Crate pallet_xcm

Pallet to handle XCM messages.

Re-exports

pub use pallet::*;

Modules

migration

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

Structs

AuthorizedAliasers

Filter for (origin: Location, target: Location) to find whether target has explicitly authorized origin to alias it.

AuthorizedAliasesEntry

EnsureResponse

EnsureOrigin implementation succeeding with a Location value to recognize and filter the Origin::Response item.

EnsureXcm

EnsureOrigin implementation succeeding with a Location value to recognize and filter the Origin::Xcm item.

IsMajorityOfBody

Filter for Location to find those which represent a strict majority approval of an identified plurality.

IsVoiceOfBody

Filter for Location to find those which represent a voice of an identified plurality.

LockTicket

ReduceTicket

TestWeightInfo

fallback implementation

UnlockTicket

XcmPassthrough

A simple passthrough where we reuse the Location-typed XCM origin as the inner value of this crate’s Origin::Xcm value.

Traits

WeightInfo

Functions

aliasers_footprint

ensure_response

Ensure that the origin o represents an XCM response origin.

ensure_xcm

Ensure that the origin o represents an XCM (Transact) origin.