pallet_xcm

Module pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

The main components of this pallet are:

Pallet, which implements all of the dispatchable extrinsics of the pallet, among other public functions.

The subset of the functions that are dispatchable can be identified either in the dispatchables module or in the Call enum.

storage_types, which contains the list of all types that are representing a storage item. Otherwise, all storage items are listed among Type Definitions.

Config, which contains the configuration trait of this pallet.

Event and Error, which are listed among the Enums.

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

migrations

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Structs

CurrentXcmVersion

An implementation of Get<u32> which just returns the latest XCM version which we can support.

GenesisConfig

Can be used to configure the genesis state of this pallet.

MaxAuthorizedAliases

The maximum number of distinct locations allowed as authorized aliases for a local origin.

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

RemoteLockedFungibleRecord

VersionDiscoveryQueueSize

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Error

The Error enum of this pallet.

Event

The Event enum of this pallet

HoldReason

A reason for this pallet placing a hold on funds.

Origin

QueryStatus

The status of a query.

VersionMigrationStage

Traits

Config

Configuration trait of this pallet.

Type Aliases

BalanceOf

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

TicketOf