Module pallet_xcm::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

execute

Execute an XCM message from a local, signed, origin.

force_default_xcm_version

Set a safe XCM version (the version that XCM should be encoded with if the most recent version a destination can accept is unknown).

force_subscribe_version_notify

Ask a location to notify us regarding their XCM version and any changes to it.

force_suspension

Set or unset the global suspension state of the XCM executor.

force_unsubscribe_version_notify

Require that a particular destination should no longer notify us regarding any XCM version changes.

force_xcm_version

Extoll that a particular destination can be communicated with through a particular version of XCM.

limited_reserve_transfer_assets

Transfer some assets from the local chain to the sovereign account of a destination chain and forward a notification XCM.

limited_teleport_assets

Teleport some assets from the local chain to some destination chain.

reserve_transfer_assets

Transfer some assets from the local chain to the sovereign account of a destination chain and forward a notification XCM.

send

Warning: Doc-Only

teleport_assets

Teleport some assets from the local chain to some destination chain.