pallet_xcm::pallet::dispatchables

Function add_authorized_alias

pub fn add_authorized_alias<T: Config>(
    aliaser: Box<VersionedLocation>,
    expires: Option<u64>,
)

Authorize another aliaser location to alias into the local origin making this call. The aliaser is only authorized until the provided expiry block number. The call can also be used for a previously authorized alias in order to update its expiry block number.

Usually useful to allow your local account to be aliased into from a remote location also under your control (like your account on another chain).

WARNING: make sure the caller origin (you) trusts the aliaser location to act in their/your name. Once authorized using this call, the aliaser can freely impersonate origin in XCM programs executed on the local chain.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::add_authorized_alias.