Function pallet_xcm::pallet::dispatchables::execute

pub fn execute<T: Config>(
    message: Box<VersionedXcm<<T as Config>::RuntimeCall>>,
    max_weight: Weight
)

Execute an XCM message from a local, signed, origin.

An event is deposited indicating whether msg could be executed completely or only partially.

No more than max_weight will be used in its attempted execution. If this is less than the maximum amount of weight that the message could take to be executed, then no execution attempt will be made.

NOTE: A successful return to this does not imply that the msg was executed successfully to completion; only that it was attempted.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::execute.