Function pallet_xcm::pallet::dispatchables::reserve_transfer_assets

pub fn reserve_transfer_assets<T: Config>(
    dest: Box<VersionedMultiLocation>,
    beneficiary: Box<VersionedMultiLocation>,
    assets: Box<VersionedMultiAssets>,
    fee_asset_item: u32
)

Transfer some assets from the local chain to the sovereign account of a destination chain and forward a notification XCM.

Fee payment on the destination side is made from the asset in the assets vector of index fee_asset_item. The weight limit for fees is not provided and thus is unlimited, with all fees taken as needed from the asset.

origin: Must be capable of withdrawing the assets and executing XCM.

dest: Destination context for the assets. Will typically be X2(Parent, Parachain(..)) to send from parachain to parachain, or X1(Parachain(..)) to send from relay to parachain.

beneficiary: A beneficiary location for the assets in the context of dest. Will generally be an AccountId32 value.

assets: The assets to be withdrawn. This should include the assets used to pay the fee on the dest side.

fee_asset_item: The index into assets of the item which should be used to pay fees.

NOTE: This function is an automatically generated, doc only, uncallable stub. designed to document the reserve_transfer_assets variant of the pallet Call enum. You should not attempt to call this function directly.