Crate pallet_xcm

Pallet to handle XCM messages.

Re-exports

pub use pallet::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

EnsureResponse

EnsureOrigin implementation succeeding with a MultiLocation value to recognize and filter the Origin::Response item.

EnsureXcm

EnsureOrigin implementation succeeding with a MultiLocation value to recognize and filter the Origin::Xcm item.

IsMajorityOfBody

Filter for MultiLocation to find those which represent a strict majority approval of an identified plurality.

XcmPassthrough

A simple passthrough where we reuse the MultiLocation-typed XCM origin as the inner value of this crate’s Origin::Xcm value.

Functions

ensure_response

Ensure that the origin o represents an XCM response origin.

ensure_xcm

Ensure that the origin o represents an XCM (Transact) origin.