pallet_xcm::pallet

Module dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

add_authorized_alias

Authorize another aliaser location to alias into the local origin making this call. The aliaser is only authorized until the provided expiry block number. The call can also be used for a previously authorized alias in order to update its expiry block number.

claim_assets

Claims assets trapped on this pallet because of leftover assets during XCM execution.

execute

Execute an XCM message from a local, signed, origin.

force_default_xcm_version

Set a safe XCM version (the version that XCM should be encoded with if the most recent version a destination can accept is unknown).

force_subscribe_version_notify

Ask a location to notify us regarding their XCM version and any changes to it.

force_suspension

Set or unset the global suspension state of the XCM executor.

force_unsubscribe_version_notify

Require that a particular destination should no longer notify us regarding any XCM version changes.

force_xcm_version

Extoll that a particular destination can be communicated with through a particular version of XCM.

limited_reserve_transfer_assets

Transfer some assets from the local chain to the destination chain through their local, destination or remote reserve.

limited_teleport_assets

Teleport some assets from the local chain to some destination chain.

remove_all_authorized_aliases

Remove all previously authorized aliasers that can alias into the local origin making this call.

remove_authorized_alias

Remove a previously authorized aliaser from the list of locations that can alias into the local origin making this call.

reserve_transfer_assets

Transfer some assets from the local chain to the destination chain through their local, destination or remote reserve.

send

Warning: Doc-Only

teleport_assets

Teleport some assets from the local chain to some destination chain.

transfer_assets

Transfer some assets from the local chain to the destination chain through their local, destination or remote reserve, or through teleports.

transfer_assets_using_type_and_then

Transfer assets from the local chain to the destination chain using explicit transfer types for assets and fees.