Module pallet_xcm::pallet::storage_types

Auto-generated docs-only module listing all defined storage types for this pallet. Note that members of this module cannot be used directly and are only provided for documentation purposes.

Structs

AssetTraps

The existing asset traps.

CurrentMigration

The current migration’s stage, if any.

LockedFungibles

Fungible assets which we know are locked on this chain.

Queries

The ongoing queries.

QueryCounter

The latest available query index.

RemoteLockedFungibles

Fungible assets which we know are locked on a remote chain.

SafeXcmVersion

Default version to encode XCM when latest version of destination is unknown. If None, then the destinations whose XCM version is unknown are considered unreachable.

SupportedVersion

The Latest versions that we know various locations support.

VersionDiscoveryQueue

Destinations whose latest XCM version we would like to know. Duplicates not allowed, and the u32 counter is the number of times that a send to the destination has been attempted, which is used as a prioritization.

VersionNotifiers

All locations that we have requested version notifications from.

VersionNotifyTargets

The target locations that are subscribed to our version changes, as well as the most recent of our versions we informed them of.