pallet_xcm::pallet
Module storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

AssetTraps

The existing asset traps.

CurrentMigration

The current migration’s stage, if any.

LockedFungibles

Fungible assets which we know are locked on this chain.

Queries

The ongoing queries.

QueryCounter

The latest available query index.

RecordedXcm

If ShouldRecordXcm is set to true, then the last XCM program executed locally will be stored here. Runtime APIs can fetch the XCM that was executed by accessing this value.

RemoteLockedFungibles

Fungible assets which we know are locked on a remote chain.

SafeXcmVersion

Default version to encode XCM when latest version of destination is unknown. If None, then the destinations whose XCM version is unknown are considered unreachable.

ShouldRecordXcm

Whether or not incoming XCMs (both executed locally and received) should be recorded. Only one XCM program will be recorded at a time. This is meant to be used in runtime APIs, and it’s advised it stays false for all other use cases, so as to not degrade regular performance.

SupportedVersion

The Latest versions that we know various locations support.

VersionDiscoveryQueue

Destinations whose latest XCM version we would like to know. Duplicates not allowed, and the u32 counter is the number of times that a send to the destination has been attempted, which is used as a prioritization.

VersionNotifiers

All locations that we have requested version notifications from.

VersionNotifyTargets

The target locations that are subscribed to our version changes, as well as the most recent of our versions we informed them of.

XcmExecutionSuspended

Global suspension state of the XCM executor.