Struct pallet_xcm::pallet::Pallet

pub struct Pallet<T>(_);

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Implementations

impl<T: Config> Pallet<T>

pub fn send( origin: OriginFor<T>, dest: Box<VersionedMultiLocation>, message: Box<VersionedXcm<()>> ) -> DispatchResult

pub fn teleport_assets( origin: OriginFor<T>, dest: Box<VersionedMultiLocation>, beneficiary: Box<VersionedMultiLocation>, assets: Box<VersionedMultiAssets>, fee_asset_item: u32 ) -> DispatchResult

Teleport some assets from the local chain to some destination chain.

Fee payment on the destination side is made from the asset in the assets vector of index fee_asset_item. The weight limit for fees is not provided and thus is unlimited, with all fees taken as needed from the asset.

origin: Must be capable of withdrawing the assets and executing XCM.

dest: Destination context for the assets. Will typically be X2(Parent, Parachain(..)) to send from parachain to parachain, or X1(Parachain(..)) to send from relay to parachain.

beneficiary: A beneficiary location for the assets in the context of dest. Will generally be an AccountId32 value.

assets: The assets to be withdrawn. The first item should be the currency used to to pay the fee on the dest side. May not be empty.

fee_asset_item: The index into assets of the item which should be used to pay fees.

pub fn reserve_transfer_assets( origin: OriginFor<T>, dest: Box<VersionedMultiLocation>, beneficiary: Box<VersionedMultiLocation>, assets: Box<VersionedMultiAssets>, fee_asset_item: u32 ) -> DispatchResult

Transfer some assets from the local chain to the sovereign account of a destination chain and forward a notification XCM.

Fee payment on the destination side is made from the asset in the assets vector of index fee_asset_item. The weight limit for fees is not provided and thus is unlimited, with all fees taken as needed from the asset.

origin: Must be capable of withdrawing the assets and executing XCM.

dest: Destination context for the assets. Will typically be X2(Parent, Parachain(..)) to send from parachain to parachain, or X1(Parachain(..)) to send from relay to parachain.

beneficiary: A beneficiary location for the assets in the context of dest. Will generally be an AccountId32 value.

assets: The assets to be withdrawn. This should include the assets used to pay the fee on the dest side.

fee_asset_item: The index into assets of the item which should be used to pay fees.

pub fn execute( origin: OriginFor<T>, message: Box<VersionedXcm<<T as SysConfig>::RuntimeCall>>, max_weight: Weight ) -> DispatchResultWithPostInfo

Execute an XCM message from a local, signed, origin.

An event is deposited indicating whether msg could be executed completely or only partially.

No more than max_weight will be used in its attempted execution. If this is less than the maximum amount of weight that the message could take to be executed, then no execution attempt will be made.

NOTE: A successful return to this does not imply that the msg was executed successfully to completion; only that some of it was executed.

pub fn force_xcm_version( origin: OriginFor<T>, location: Box<MultiLocation>, version: XcmVersion ) -> DispatchResult

Extoll that a particular destination can be communicated with through a particular version of XCM.

origin: Must be an origin specified by AdminOrigin.

location: The destination that is being described.

xcm_version: The latest version of XCM that location supports.

pub fn force_default_xcm_version( origin: OriginFor<T>, maybe_xcm_version: Option<XcmVersion> ) -> DispatchResult

Set a safe XCM version (the version that XCM should be encoded with if the most recent version a destination can accept is unknown).

origin: Must be an origin specified by AdminOrigin.

maybe_xcm_version: The default XCM encoding version, or None to disable.

pub fn force_subscribe_version_notify( origin: OriginFor<T>, location: Box<VersionedMultiLocation> ) -> DispatchResult

Ask a location to notify us regarding their XCM version and any changes to it.

origin: Must be an origin specified by AdminOrigin.

location: The location to which we should subscribe for XCM version notifications.

pub fn force_unsubscribe_version_notify( origin: OriginFor<T>, location: Box<VersionedMultiLocation> ) -> DispatchResult

Require that a particular destination should no longer notify us regarding any XCM version changes.

origin: Must be an origin specified by AdminOrigin.

location: The location to which we are currently subscribed for XCM version notifications which we no longer desire.

pub fn limited_reserve_transfer_assets( origin: OriginFor<T>, dest: Box<VersionedMultiLocation>, beneficiary: Box<VersionedMultiLocation>, assets: Box<VersionedMultiAssets>, fee_asset_item: u32, weight_limit: WeightLimit ) -> DispatchResult

Transfer some assets from the local chain to the sovereign account of a destination chain and forward a notification XCM.

Fee payment on the destination side is made from the asset in the assets vector of index fee_asset_item, up to enough to pay for weight_limit of weight. If more weight is needed than weight_limit, then the operation will fail and the assets send may be at risk.

origin: Must be capable of withdrawing the assets and executing XCM.

dest: Destination context for the assets. Will typically be X2(Parent, Parachain(..)) to send from parachain to parachain, or X1(Parachain(..)) to send from relay to parachain.

beneficiary: A beneficiary location for the assets in the context of dest. Will generally be an AccountId32 value.

assets: The assets to be withdrawn. This should include the assets used to pay the fee on the dest side.

fee_asset_item: The index into assets of the item which should be used to pay fees.

weight_limit: The remote-side weight limit, if any, for the XCM fee purchase.

pub fn limited_teleport_assets( origin: OriginFor<T>, dest: Box<VersionedMultiLocation>, beneficiary: Box<VersionedMultiLocation>, assets: Box<VersionedMultiAssets>, fee_asset_item: u32, weight_limit: WeightLimit ) -> DispatchResult

Teleport some assets from the local chain to some destination chain.

Fee payment on the destination side is made from the asset in the assets vector of index fee_asset_item, up to enough to pay for weight_limit of weight. If more weight is needed than weight_limit, then the operation will fail and the assets send may be at risk.

origin: Must be capable of withdrawing the assets and executing XCM.

dest: Destination context for the assets. Will typically be X2(Parent, Parachain(..)) to send from parachain to parachain, or X1(Parachain(..)) to send from relay to parachain.

beneficiary: A beneficiary location for the assets in the context of dest. Will generally be an AccountId32 value.

assets: The assets to be withdrawn. The first item should be the currency used to to pay the fee on the dest side. May not be empty.

fee_asset_item: The index into assets of the item which should be used to pay fees.

weight_limit: The remote-side weight limit, if any, for the XCM fee purchase.

pub fn force_suspension(origin: OriginFor<T>, suspended: bool) -> DispatchResult

Set or unset the global suspension state of the XCM executor.

origin: Must be an origin specified by AdminOrigin.

suspended: true to suspend, false to resume.

impl<T: Config> Pallet<T>

pub fn query<KArg>(k: KArg) -> Option<QueryStatus<BlockNumberFor<T>>>where KArg: EncodeLike<QueryId>,

“ The ongoing queries.“

impl<T: Config> Pallet<T>

pub fn asset_trap<KArg>(k: KArg) -> u32where KArg: EncodeLike<H256>,

“ The existing asset traps.“ “” “ Key is the blake2 256 hash of (origin, versioned MultiAssets) pair. Value is the number of“ “ times this pair has been trapped (usually just 1 if it exists at all).“

impl<T: Config> Pallet<T>

pub fn request_version_notify(dest: impl Into<MultiLocation>) -> XcmResult

Request that dest informs us of its version.

pub fn unrequest_version_notify(dest: impl Into<MultiLocation>) -> XcmResult

Request that dest ceases informing us of its version.

pub fn send_xcm( interior: impl Into<Junctions>, dest: impl Into<MultiLocation>, message: Xcm<()> ) -> Result<XcmHash, SendError>

Relay an XCM message from a given interior location in this context to a given dest location. The fee_payer is charged for the delivery unless None in which case fees are not charged (and instead borne by the chain).

pub fn check_account() -> T::AccountId

pub fn report_outcome_notify( message: &mut Xcm<()>, responder: impl Into<MultiLocation>, notify: impl Into<<T as Config>::RuntimeCall>, timeout: BlockNumberFor<T> ) -> Result<(), XcmError>

Consume message and return another which is equivalent to it except that it reports back the outcome and dispatches notify on this chain.

message: The message whose outcome should be reported.

responder: The origin from which a response should be expected.

notify: A dispatchable function which will be called once the outcome of message is known. It may be a dispatchable in any pallet of the local chain, but other than the usual origin, it must accept exactly two arguments: query_id: QueryId and outcome: Response, and in that order. It should expect that the origin is Origin::Response and will contain the responder’s location.

timeout: The block number after which it is permissible for notify not to be called even if a response is received.

report_outcome_notify may return an error if the responder is not invertible.

It is assumed that the querier of the response will be Here.

NOTE: notify gets called as part of handling an incoming message, so it should be lightweight. Its weight is estimated during this function and stored ready for weighing ReportOutcome on the way back. If it turns out to be heavier once it returns then reporting the outcome will fail. Futhermore if the estimate is too high, then it may be put in the overweight queue and need to be manually executed.

pub fn new_notify_query( responder: impl Into<MultiLocation>, notify: impl Into<<T as Config>::RuntimeCall>, timeout: BlockNumberFor<T>, match_querier: impl Into<MultiLocation> ) -> u64

Attempt to create a new query ID and register it as a query that is yet to respond, and which will call a dispatchable when a response happens.