Trait pallet_xcm::pallet::Config

pub trait Config: Config {

    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type SendXcmOrigin: EnsureOrigin<<Self as SysConfig>::Origin, Success = MultiLocation>;
    type XcmRouter: SendXcm;
    type ExecuteXcmOrigin: EnsureOrigin<<Self as SysConfig>::Origin, Success = MultiLocation>;
    type XcmExecuteFilter: Contains<(MultiLocation, Xcm<<Self as SysConfig>::Call>)>;
    type XcmExecutor: ExecuteXcm<<Self as SysConfig>::Call>;
    type XcmTeleportFilter: Contains<(MultiLocation, Vec<MultiAsset>)>;
    type XcmReserveTransferFilter: Contains<(MultiLocation, Vec<MultiAsset>)>;
    type Weigher: WeightBounds<<Self as SysConfig>::Call>;
    type LocationInverter: InvertLocation;
    type Origin: From<Origin> + From<<Self as SysConfig>::Origin>;
    type Call: Parameter + GetDispatchInfo + IsType<<Self as Config>::Call> + Dispatchable<Origin = <Self as Config>::Origin, PostInfo = PostDispatchInfo>;
    type AdvertisedXcmVersion: Get<XcmVersion>;

    const VERSION_DISCOVERY_QUEUE_SIZE: u32;

}

The module configuration trait.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

The overarching event type.

type SendXcmOrigin: EnsureOrigin<<Self as SysConfig>::Origin, Success = MultiLocation>

Required origin for sending XCM messages. If successful, it resolves to MultiLocation which exists as an interior location within this chain’s XCM context.

type XcmRouter: SendXcm

The type used to actually dispatch an XCM to its destination.

type ExecuteXcmOrigin: EnsureOrigin<<Self as SysConfig>::Origin, Success = MultiLocation>

Required origin for executing XCM messages, including the teleport functionality. If successful, then it resolves to MultiLocation which exists as an interior location within this chain’s XCM context.

type XcmExecuteFilter: Contains<(MultiLocation, Xcm<<Self as SysConfig>::Call>)>

Our XCM filter which messages to be executed using XcmExecutor must pass.

type XcmExecutor: ExecuteXcm<<Self as SysConfig>::Call>

Something to execute an XCM message.

type XcmTeleportFilter: Contains<(MultiLocation, Vec<MultiAsset>)>

Our XCM filter which messages to be teleported using the dedicated extrinsic must pass.

type XcmReserveTransferFilter: Contains<(MultiLocation, Vec<MultiAsset>)>

Our XCM filter which messages to be reserve-transferred using the dedicated extrinsic must pass.

type Weigher: WeightBounds<<Self as SysConfig>::Call>

Means of measuring the weight consumed by an XCM message locally.

type LocationInverter: InvertLocation

Means of inverting a location.

type Origin: From<Origin> + From<<Self as SysConfig>::Origin>

The outer Origin type.

type Call: Parameter + GetDispatchInfo + IsType<<Self as Config>::Call> + Dispatchable<Origin = <Self as Config>::Origin, PostInfo = PostDispatchInfo>

The outer Call type.

type AdvertisedXcmVersion: Get<XcmVersion>

The latest supported version that we advertise. Generally just set it to pallet_xcm::CurrentXcmVersion.

Required Associated Constants

const VERSION_DISCOVERY_QUEUE_SIZE: u32