Trait pallet_xcm::pallet::Config

pub trait Config: Config {

    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type Currency: LockableCurrency<Self::AccountId, Moment = Self::BlockNumber>;
    type CurrencyMatcher: MatchesFungible<BalanceOf<Self>>;
    type SendXcmOrigin: EnsureOrigin<<Self as SysConfig>::RuntimeOrigin, Success = MultiLocation>;
    type XcmRouter: SendXcm;
    type ExecuteXcmOrigin: EnsureOrigin<<Self as SysConfig>::RuntimeOrigin, Success = MultiLocation>;
    type XcmExecuteFilter: Contains<(MultiLocation, Xcm<<Self as SysConfig>::RuntimeCall>)>;
    type XcmExecutor: ExecuteXcm<<Self as SysConfig>::RuntimeCall>;
    type XcmTeleportFilter: Contains<(MultiLocation, Vec<MultiAsset>)>;
    type XcmReserveTransferFilter: Contains<(MultiLocation, Vec<MultiAsset>)>;
    type Weigher: WeightBounds<<Self as SysConfig>::RuntimeCall>;
    type UniversalLocation: Get<InteriorMultiLocation>;
    type RuntimeOrigin: From<Origin> + From<<Self as SysConfig>::RuntimeOrigin>;
    type RuntimeCall: Parameter + GetDispatchInfo + IsType<<Self as Config>::RuntimeCall> + Dispatchable<RuntimeOrigin = <Self as Config>::RuntimeOrigin, PostInfo = PostDispatchInfo>;
    type AdvertisedXcmVersion: Get<XcmVersion>;
    type AdminOrigin: EnsureOrigin<<Self as SysConfig>::RuntimeOrigin>;
    type TrustedLockers: ContainsPair<MultiLocation, MultiAsset>;
    type SovereignAccountOf: Convert<MultiLocation, Self::AccountId>;
    type MaxLockers: Get<u32>;
    type WeightInfo: WeightInfo;

    const VERSION_DISCOVERY_QUEUE_SIZE: u32;

}

The module configuration trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

The overarching event type.

type Currency: LockableCurrency<Self::AccountId, Moment = Self::BlockNumber>

A lockable currency.

type CurrencyMatcher: MatchesFungible<BalanceOf<Self>>

The MultiAsset matcher for Currency.

type SendXcmOrigin: EnsureOrigin<<Self as SysConfig>::RuntimeOrigin, Success = MultiLocation>

Required origin for sending XCM messages. If successful, it resolves to MultiLocation which exists as an interior location within this chain’s XCM context.

type XcmRouter: SendXcm

The type used to actually dispatch an XCM to its destination.

type ExecuteXcmOrigin: EnsureOrigin<<Self as SysConfig>::RuntimeOrigin, Success = MultiLocation>

Required origin for executing XCM messages, including the teleport functionality. If successful, then it resolves to MultiLocation which exists as an interior location within this chain’s XCM context.

type XcmExecuteFilter: Contains<(MultiLocation, Xcm<<Self as SysConfig>::RuntimeCall>)>

Our XCM filter which messages to be executed using XcmExecutor must pass.

type XcmExecutor: ExecuteXcm<<Self as SysConfig>::RuntimeCall>

Something to execute an XCM message.

type XcmTeleportFilter: Contains<(MultiLocation, Vec<MultiAsset>)>

Our XCM filter which messages to be teleported using the dedicated extrinsic must pass.

type XcmReserveTransferFilter: Contains<(MultiLocation, Vec<MultiAsset>)>

Our XCM filter which messages to be reserve-transferred using the dedicated extrinsic must pass.

type Weigher: WeightBounds<<Self as SysConfig>::RuntimeCall>

Means of measuring the weight consumed by an XCM message locally.

type UniversalLocation: Get<InteriorMultiLocation>

This chain’s Universal Location.

type RuntimeOrigin: From<Origin> + From<<Self as SysConfig>::RuntimeOrigin>

The runtime Origin type.

type RuntimeCall: Parameter + GetDispatchInfo + IsType<<Self as Config>::RuntimeCall> + Dispatchable<RuntimeOrigin = <Self as Config>::RuntimeOrigin, PostInfo = PostDispatchInfo>

The runtime Call type.

type AdvertisedXcmVersion: Get<XcmVersion>

The latest supported version that we advertise. Generally just set it to pallet_xcm::CurrentXcmVersion.

type AdminOrigin: EnsureOrigin<<Self as SysConfig>::RuntimeOrigin>

The origin that is allowed to call privileged operations on the XCM pallet

type TrustedLockers: ContainsPair<MultiLocation, MultiAsset>

The assets which we consider a given origin is trusted if they claim to have placed a lock.

type SovereignAccountOf: Convert<MultiLocation, Self::AccountId>

How to get an AccountId value from a MultiLocation, useful for handling asset locks.

type MaxLockers: Get<u32>

The maximum number of local XCM locks that a single account may have.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

Required Associated Constants

const VERSION_DISCOVERY_QUEUE_SIZE: u32