pallet_xcm::pallet

Trait Config

pub trait Config: Config + Config<RuntimeEvent: From<Event<Self>>> {

    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type Currency: LockableCurrency<Self::AccountId, Moment = BlockNumberFor<Self>>;
    type CurrencyMatcher: MatchesFungible<BalanceOf<Self>>;
    type AuthorizedAliasConsideration: Consideration<Self::AccountId, Footprint>;
    type SendXcmOrigin: EnsureOrigin<<Self as SysConfig>::RuntimeOrigin, Success = Location>;
    type XcmRouter: SendXcm;
    type ExecuteXcmOrigin: EnsureOrigin<<Self as SysConfig>::RuntimeOrigin, Success = Location>;
    type XcmExecuteFilter: Contains<(Location, Xcm<<Self as Config>::RuntimeCall>)>;
    type XcmExecutor: ExecuteXcm<<Self as Config>::RuntimeCall> + XcmAssetTransfers + FeeManager;
    type XcmTeleportFilter: Contains<(Location, Vec<Asset>)>;
    type XcmReserveTransferFilter: Contains<(Location, Vec<Asset>)>;
    type Weigher: WeightBounds<<Self as Config>::RuntimeCall>;
    type UniversalLocation: Get<InteriorLocation>;
    type RuntimeOrigin: From<Origin> + From<<Self as SysConfig>::RuntimeOrigin>;
    type RuntimeCall: Parameter + GetDispatchInfo + Dispatchable<RuntimeOrigin = <Self as Config>::RuntimeOrigin, PostInfo = PostDispatchInfo>;
    type AdvertisedXcmVersion: Get<XcmVersion>;
    type AdminOrigin: EnsureOrigin<<Self as SysConfig>::RuntimeOrigin>;
    type TrustedLockers: ContainsPair<Location, Asset>;
    type SovereignAccountOf: ConvertLocation<Self::AccountId>;
    type MaxLockers: Get<u32>;
    type MaxRemoteLockConsumers: Get<u32>;
    type RemoteLockConsumerIdentifier: Parameter + Member + MaxEncodedLen + Ord + Copy;
    type WeightInfo: WeightInfo;

    const VERSION_DISCOVERY_QUEUE_SIZE: u32;

}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait. The module configuration trait.

Required Associated Constants

const VERSION_DISCOVERY_QUEUE_SIZE: u32

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

The overarching event type.

type Currency: LockableCurrency<Self::AccountId, Moment = BlockNumberFor<Self>>

A lockable currency.

type CurrencyMatcher: MatchesFungible<BalanceOf<Self>>

The Asset matcher for Currency.

type AuthorizedAliasConsideration: Consideration<Self::AccountId, Footprint>

A means of providing some cost while Authorized Aliasers data is stored on-chain.

type SendXcmOrigin: EnsureOrigin<<Self as SysConfig>::RuntimeOrigin, Success = Location>

Required origin for sending XCM messages. If successful, it resolves to Location which exists as an interior location within this chain’s XCM context.

type XcmRouter: SendXcm

The type used to actually dispatch an XCM to its destination.

type ExecuteXcmOrigin: EnsureOrigin<<Self as SysConfig>::RuntimeOrigin, Success = Location>

Required origin for executing XCM messages, including the teleport functionality. If successful, then it resolves to Location which exists as an interior location within this chain’s XCM context.

type XcmExecuteFilter: Contains<(Location, Xcm<<Self as Config>::RuntimeCall>)>

Our XCM filter which messages to be executed using XcmExecutor must pass.

type XcmExecutor: ExecuteXcm<<Self as Config>::RuntimeCall> + XcmAssetTransfers + FeeManager

Something to execute an XCM message.

type XcmTeleportFilter: Contains<(Location, Vec<Asset>)>

Our XCM filter which messages to be teleported using the dedicated extrinsic must pass.

type XcmReserveTransferFilter: Contains<(Location, Vec<Asset>)>

Our XCM filter which messages to be reserve-transferred using the dedicated extrinsic must pass.

type Weigher: WeightBounds<<Self as Config>::RuntimeCall>

Means of measuring the weight consumed by an XCM message locally.

type UniversalLocation: Get<InteriorLocation>

This chain’s Universal Location.

type RuntimeOrigin: From<Origin> + From<<Self as SysConfig>::RuntimeOrigin>

The runtime Origin type.

type RuntimeCall: Parameter + GetDispatchInfo + Dispatchable<RuntimeOrigin = <Self as Config>::RuntimeOrigin, PostInfo = PostDispatchInfo>

The runtime Call type.

type AdvertisedXcmVersion: Get<XcmVersion>

The latest supported version that we advertise. Generally just set it to pallet_xcm::CurrentXcmVersion.

type AdminOrigin: EnsureOrigin<<Self as SysConfig>::RuntimeOrigin>

The origin that is allowed to call privileged operations on the XCM pallet

type TrustedLockers: ContainsPair<Location, Asset>

The assets which we consider a given origin is trusted if they claim to have placed a lock.

type SovereignAccountOf: ConvertLocation<Self::AccountId>

How to get an AccountId value from a Location, useful for handling asset locks.

type MaxLockers: Get<u32>

The maximum number of local XCM locks that a single account may have.

type MaxRemoteLockConsumers: Get<u32>

The maximum number of consumers a single remote lock may have.

type RemoteLockConsumerIdentifier: Parameter + Member + MaxEncodedLen + Ord + Copy

The ID type for local consumers of remote locks.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.