Trait pallet_xcm::pallet::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<Self::Event>;
    type SendXcmOrigin: EnsureOrigin<Self::Origin, Success = MultiLocation>;
    type XcmRouter: SendXcm;
    type ExecuteXcmOrigin: EnsureOrigin<Self::Origin, Success = MultiLocation>;
    type XcmExecuteFilter: Contains<(MultiLocation, Xcm<Self::Call>)>;
    type XcmExecutor: ExecuteXcm<Self::Call>;
    type XcmTeleportFilter: Contains<(MultiLocation, Vec<MultiAsset>)>;
    type XcmReserveTransferFilter: Contains<(MultiLocation, Vec<MultiAsset>)>;
    type Weigher: WeightBounds<Self::Call>;
    type LocationInverter: InvertLocation;
    type Origin: From<Origin> + From<Self::Origin>;
    type Call: Parameter + GetDispatchInfo + IsType<Self::Call> + Dispatchable<Origin = Self::Origin, PostInfo = PostDispatchInfo>;
}

The module configuration trait.

Associated Types

type Event: From<Event<Self>> + IsType<Self::Event>

The overarching event type.

type SendXcmOrigin: EnsureOrigin<Self::Origin, Success = MultiLocation>

Required origin for sending XCM messages. If successful, the it resolves to MultiLocation which exists as an interior location within this chain’s XCM context.

type XcmRouter: SendXcm

The type used to actually dispatch an XCM to its destination.

type ExecuteXcmOrigin: EnsureOrigin<Self::Origin, Success = MultiLocation>

Required origin for executing XCM messages, including the teleport functionality. If successful, then it resolves to MultiLocation which exists as an interior location within this chain’s XCM context.

type XcmExecuteFilter: Contains<(MultiLocation, Xcm<Self::Call>)>

Our XCM filter which messages to be executed using XcmExecutor must pass.

type XcmExecutor: ExecuteXcm<Self::Call>

Something to execute an XCM message.

type XcmTeleportFilter: Contains<(MultiLocation, Vec<MultiAsset>)>

Our XCM filter which messages to be teleported using the dedicated extrinsic must pass.

type XcmReserveTransferFilter: Contains<(MultiLocation, Vec<MultiAsset>)>

Our XCM filter which messages to be reserve-transferred using the dedicated extrinsic must pass.

type Weigher: WeightBounds<Self::Call>

Means of measuring the weight consumed by an XCM message locally.

type LocationInverter: InvertLocation

Means of inverting a location.

type Origin: From<Origin> + From<Self::Origin>

The outer Origin type.

type Call: Parameter + GetDispatchInfo + IsType<Self::Call> + Dispatchable<Origin = Self::Origin, PostInfo = PostDispatchInfo>

The outer Call type.

Loading content...